Noticias_live/requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.0.8
django-ckeditor==5.9.0
django-js-asset==1.2.2
gunicorn==20.0.4
Pillow==7.2.0
python-decouple==3.3
pytz==2020.1
sqlparse==0.3.1
